Revert tokenizer consolidation into utils
